refactor(cli): extract exception classification to _lifecycle.py
Creates classify_exception() and EXIT_CODE_BY_STATUS in _lifecycle.py,
eliminating the duplicated status→exit-code dict from agent.py and the
per-file except-chain copies in fanout.py and flow.py. Updates all test
import sites to reference the canonical locations.
